deployed_llms/colpali_cloud/requirements.txt
tqdm
torch
pydantic
pillow
fastapi
requests
